backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.0
PyMuPDF>=1.23
scikit-learn>=1.3
requests>=2.31
social-auth-app-django>=5.4.0
social-auth-core>=4.5.0
pymongo>=4.6
gunicorn>=21.0
whitenoise>=6.6